LPS-185164 Move configuration provider to manager in order to make sense to add more methods
