+++ modules/apps/site/site-admin-web/src/main/java/com/liferay/site/admin/web/internal/display/context/MenuAccessConfigurationDisplayContext.java
@@ -35,7 +35,7 @@ import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portlet.rolesadmin.search.RoleSearch;
 import com.liferay.roles.item.selector.regular.role.RegularRoleItemSelectorCriterion;
 import com.liferay.roles.item.selector.site.role.SiteRoleItemSelectorCriterion;
-import com.liferay.site.configuration.MenuAccessConfigurationProvider;
+import com.liferay.site.configuration.manager.MenuAccessConfigurationManager;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -54,12 +54,12 @@ public class MenuAccessConfigurationDisplayContext {
 
 	public MenuAccessConfigurationDisplayContext(
 		HttpServletRequest httpServletRequest, ItemSelector itemSelector,
-		MenuAccessConfigurationProvider menuAccessConfigurationProvider,
+		MenuAccessConfigurationManager menuAccessConfigurationManager,
 		Portal portal, RoleLocalService roleLocalService) {
 
 		_httpServletRequest = httpServletRequest;
 		_itemSelector = itemSelector;
-		_menuAccessConfigurationProvider = menuAccessConfigurationProvider;
+		_menuAccessConfigurationManager = menuAccessConfigurationManager;
 		_roleLocalService = roleLocalService;
 
 		_liferayPortletRequest = portal.getLiferayPortletRequest(
@@ -88,7 +88,7 @@ public class MenuAccessConfigurationDisplayContext {
 		itemSelectorCriteria.add(regularRoleItemSelectorCriterion);
 
 		String[] roleIds =
-			_menuAccessConfigurationProvider.getAccessToControlMenuRoleIds(
+			_menuAccessConfigurationManager.getAccessToControlMenuRoleIds(
 				_themeDisplay.getScopeGroupId());
 
 		long[] checkedRoleIds = new long[roleIds.length];
@@ -143,7 +143,7 @@ public class MenuAccessConfigurationDisplayContext {
 		List<Role> roles = new ArrayList<>();
 
 		String[] accessToControlMenuRoleIds =
-			_menuAccessConfigurationProvider.getAccessToControlMenuRoleIds(
+			_menuAccessConfigurationManager.getAccessToControlMenuRoleIds(
 				_themeDisplay.getScopeGroupId());
 
 		for (String roleId : accessToControlMenuRoleIds) {
@@ -180,7 +180,7 @@ public class MenuAccessConfigurationDisplayContext {
 	}
 
 	public boolean isShowControlMenuByRole() throws Exception {
-		return _menuAccessConfigurationProvider.isShowControlMenuByRole(
+		return _menuAccessConfigurationManager.isShowControlMenuByRole(
 			_themeDisplay.getScopeGroupId());
 	}
 
@@ -202,8 +202,8 @@ public class MenuAccessConfigurationDisplayContext {
 	private final ItemSelector _itemSelector;
 	private final LiferayPortletRequest _liferayPortletRequest;
 	private final LiferayPortletResponse _liferayPortletResponse;
-	private final MenuAccessConfigurationProvider
-		_menuAccessConfigurationProvider;
+	private final MenuAccessConfigurationManager
+		_menuAccessConfigurationManager;
 	private final RoleLocalService _roleLocalService;
 	private final ThemeDisplay _themeDisplay;
