+++ modules/apps/site/site-api/bnd.bnd
@@ -3,6 +3,7 @@ Bundle-SymbolicName: com.liferay.site.api
 Bundle-Version: 16.5.1
 Export-Package:\
 	com.liferay.site.configuration,\
+	com.liferay.site.configuration.manager,\
 	com.liferay.site.constants,\
 	com.liferay.site.display.context,\
 	com.liferay.site.exception,\
